requirements.txt
distlib==0.3.6
filelock==3.8.0
numpy==1.23.3
Pillow==9.2.0
platformdirs==2.5.2
virtualenv==20.16.5
